requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-community
langchain-huggingface
huggingface_hub
requests
python-dotenv
pydantic
